+++ src/utils/datastore/index.js
@@ -1,4 +1,4 @@
-import about from '@/utils/datastore/about'
+import about, { plugin as aboutPlugin } from '@/utils/datastore/about'
 import connectivity from '@/utils/datastore/connectivity'
 import presence from '@/utils/datastore/presence'
 import refresh from '@/utils/datastore/refresh'
@@ -13,5 +13,6 @@ export default {
   },
   plugins: [
     refreshPlugin,
+    aboutPlugin,
   ],
 }