+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Please document your changes in this format:
 - Subscribe everybody to conflict resolution emails by default @nicksellen
 - Speed up page load (hopefully) @tiltec @nicksellen
 - Hide buttons in group sidenav, change notification icons, added settings to latest messages menu [ #1987 @brnsolikyl @nicksellen
+- Show instruction on group preview page how to revive groups without members @tiltec
 
 ### Fixed
 - Unsubscribe link in pickup notification email did not work @nicksellen