+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Please document your changes in this format:
 - Unsubscribe link in pickup notification email did not work @nicksellen
 - Names in application chat were normally not visible for the applicant @tiltec
 - "Forgot your password?" page didn't show validation errors @tiltec
+- Selecting to join playground group after signup didn't actually join the group @tiltec
 
 ## [8.4.0] - 2020-04-25
 ### Added